Struct wasmtime_cache::CacheConfig

pub struct CacheConfig { /* private fields */ }

Global configuration for how the cache is managed

Implementations

impl CacheConfig

pub fn worker_event_queue_size(&self) -> u64

Returns $setting.

Panics if the cache is disabled.

pub fn baseline_compression_level(&self) -> i32

Returns $setting.

Panics if the cache is disabled.

pub fn optimized_compression_level(&self) -> i32

Returns $setting.

Panics if the cache is disabled.

pub fn optimized_compression_usage_counter_threshold(&self) -> u64

Returns $setting.

Panics if the cache is disabled.

pub fn cleanup_interval(&self) -> Duration

Returns $setting.

Panics if the cache is disabled.

pub fn optimizing_compression_task_timeout(&self) -> Duration

Returns $setting.

Panics if the cache is disabled.

pub fn allowed_clock_drift_for_files_from_future(&self) -> Duration

Returns $setting.

Panics if the cache is disabled.

pub fn file_count_soft_limit(&self) -> u64

Returns $setting.

Panics if the cache is disabled.

pub fn files_total_size_soft_limit(&self) -> u64

Returns $setting.

Panics if the cache is disabled.

pub fn file_count_limit_percent_if_deleting(&self) -> u8

Returns $setting.

Panics if the cache is disabled.

pub fn files_total_size_limit_percent_if_deleting(&self) -> u8

Returns $setting.

Panics if the cache is disabled.

pub fn enabled(&self) -> bool

Returns true if and only if the cache is enabled.

pub fn directory(&self) -> &PathBuf

Returns path to the cache directory.

Panics if the cache is disabled.

pub fn new_cache_disabled() -> Self

Creates a new set of configuration which represents a disabled cache

pub fn from_file(config_file: Option<&Path>) -> Result<Self>

Parses cache configuration from the file specified

pub fn cache_hits(&self) -> usize

Returns the number of cache hits seen so far

pub fn cache_misses(&self) -> usize

Returns the number of cache misses seen so far